need to manage concurrent updates with amended forms
- there can be multiple instances of amended forms for a
give CIK/quarterly filing.  When processing asynchronously
they can interfere with each other. This results in DB exceptions
due to duplicate keys.
Manage this by queueing on the CIK/quarter end date.  Take advantage
of the fact that inserting into a list either succeeds or fails
atomically. Put this function into a class and guard the list with
a mutex.